some bug fixes in gulpfile.js and apprunner.js is removed
